Fixed a bug with the sprite loader and various things related to the scene switching code. Now the intro works correctly with a minor graphics glitch.
svn-id: r19588

+++ kyra/script.h
@@ -70,7 +70,7 @@ protected:
 	
 	KyraEngine *_vm;
 	ScriptState *_curScript;
-	uint32 _parameter;
+	int16 _parameter;
 	bool _continue;
 
 	typedef void (ScriptHelper::*CommandProc)();
